requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cssselect==1.1.0
cycler==0.10.0
fake-useragent==0.1.11
filelock==3.0.12
futures==3.1.1
goslate==1.5.1
graphviz==0.17
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
Levenshtein==0.13.0
lxml==4.6.3
matplotlib==3.4.3
mysql-connector==2.2.9
nltk==3.6.2
numpy==1.21.0
pandas==1.3.0
parse==1.19.0
Pillow==8.3.1
publicsuffix==1.1.1
pyasn==1.6.1
pyee==8.1.0
pyenchant==3.2.1
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
pytesseract==0.3.8
python-dateutil==2.8.1
pytz==2021.1
regex==2021.8.3
requests==2.25.1
requests-file==1.5.1
requests-html==0.10.0
scikit-learn==0.24.2
scipy==1.7.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
threadpoolctl==2.1.0
tldextract==3.1.0
tqdm==4.61.2
tranco==0.6
Unidecode==1.2.0
urllib3==1.26.6
utils==1.0.1
virustotal-python==0.1.3
w3lib==1.22.0
websockets==8.1
wordninja==2.0.0
